third_party/WebKit/LayoutTests/web-animations-api/timed-item-expected.txt

This is a testharness.js-based test.
FAIL localTime and currentIteration are null when the KeyframeEffect is not associated with an Animation assert_equals: expected (object) null but got (undefined) undefined
PASS TimedItem.localTime and TimedItem.currentIteration return reasonable values when an keyframeEffect is in effect 
FAIL TimedItem.currentIteration is null when keyframeEffect is not in effect assert_equals: expected (object) null but got (undefined) undefined
PASS TimedItem startTime, endTime, duration, activeDuration are sensible for a simple keyframeEffect 
PASS TimedItem startTime, endTime, duration, activeDuration are sensible for keyframeEffects with delays and iterations 
PASS TimedItem duration is calculated when no duration is specified 
PASS Restricted double attributes on the Timing interface throws for non-finite values. 
Harness: the test ran to completion.
